config.xml: add 'store' flag to image types.
Flag 'false' for all image types except 'medium_default'.

This was simply missing, despite this image type being used in
stores.tpl. Nevertheless it often worked, because the thirty bees
installer defined such a flag. Which is gone, if one switches to
another theme and back to this theme, of course.

Also line breaks to keep 80 characters width.